Extract transaction.proto from daml_kvutils.proto
CHANGELOG_BEGIN

CHANGELOG_END
