+++ ledger/participant-state/kvutils/src/main/protobuf/com/daml/ledger/participant/state/kvutils/wire/submission.proto
@@ -10,6 +10,7 @@ option csharp_namespace = "Com.Daml.Ledger.Participant.State.KVUtils.Wire";
 import "com/daml/ledger/participant/state/kvutils/daml_kvutils.proto";
 import "com/daml/ledger/participant/state/kvutils/store/state.proto";
 import "com/daml/ledger/participant/state/kvutils/store/events/package_upload.proto";
+import "com/daml/ledger/participant/state/kvutils/store/events/transaction.proto";
 import "com/daml/ledger/participant/state/kvutils/wire/configuration_submission.proto";
 
 // A submission to the ledger: a payload and its inputs if any.
@@ -18,7 +19,7 @@ import "com/daml/ledger/participant/state/kvutils/wire/configuration_submission.
 message DamlSubmission {
   repeated com.daml.ledger.participant.state.kvutils.store.DamlStateKey input_daml_state = 1;
   oneof payload {
-    DamlTransactionEntry transaction_entry = 2;
+    com.daml.ledger.participant.state.kvutils.store.events.DamlTransactionEntry transaction_entry = 2;
     com.daml.ledger.participant.state.kvutils.store.events.DamlPackageUploadEntry package_upload_entry = 3;
     DamlConfigurationSubmission configuration_submission = 4;
     DamlPartyAllocationEntry party_allocation_entry = 5;
