+++ ledger/participant-state/kvutils/src/test/suite/scala/com/daml/ledger/participant/state/kvutils/ConversionsSpec.scala
@@ -8,13 +8,18 @@ import java.time.{Duration, Instant}
 import com.daml.ledger.api.DeduplicationPeriod
 import com.daml.ledger.configuration.LedgerTimeModel
 import com.daml.ledger.participant.state.kvutils.Conversions._
-import com.daml.ledger.participant.state.kvutils.DamlKvutils.DamlTransactionBlindingInfo.{
-  DisclosureEntry,
-  DivulgenceEntry,
-}
 import com.daml.ledger.participant.state.kvutils.DamlKvutils._
 import com.daml.ledger.participant.state.kvutils.committer.transaction.Rejection
 import com.daml.ledger.participant.state.kvutils.store.DamlStateKey
+import com.daml.ledger.participant.state.kvutils.store.events.DamlTransactionBlindingInfo.{
+  DisclosureEntry,
+  DivulgenceEntry,
+}
+import com.daml.ledger.participant.state.kvutils.store.events.{
+  DamlSubmitterInfo,
+  DamlTransactionBlindingInfo,
+  DamlTransactionRejectionEntry,
+}
 import com.daml.ledger.participant.state.v2.Update.CommandRejected
 import com.daml.lf.crypto
 import com.daml.lf.crypto.Hash
