+++ ledger/participant-state/kvutils/src/test/suite/scala/com/daml/ledger/validator/batch/ConflictDetectionSpec.scala
@@ -4,7 +4,7 @@
 package com.daml.ledger.validator.batch
 
 import com.codahale.metrics.MetricRegistry
-import com.daml.ledger.participant.state.kvutils.DamlKvutils._
+import com.daml.ledger.participant.state.kvutils.store.events.DamlTransactionRejectionEntry
 import com.daml.ledger.participant.state.kvutils.store.events.PackageUpload.DamlPackageUploadEntry
 import com.daml.ledger.participant.state.kvutils.store.{
   DamlContractKey,
